requirements.txt
torchmetrics==1.5.2
scikit-learn==1.3.2
tensorboard==2.14.0
tensorboardx==2.6.2.2
timm==1.0.7
numpy==1.24.3
opencv-python==4.9.0.80
pycocotools
pandas